# docs/api/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Samsara Engine API",
    "description": "Engineering foundation only; traditional production rules remain disabled.",
    "version": "0.1.0-sprint1a"
  },
  "paths": {
...
    "/api/v1/auth/invitations/accept": {
      "post": {
        "summary": "Accept Invitation",
        "operationId": "accept_invitation_api_v1_auth_invitations_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationAccept"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "Logout",
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "name": "__Host-session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_api_v1_me_get",
        "parameters": [
          {
            "name": "__Host-session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles": {
      "post": {
        "summary": "Create Profile",
        "operationId": "create_profile_api_v1_profiles_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "__Host-session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles/{profile_id}": {
      "get": {
        "summary": "Get Profile",
        "operationId": "get_profile_api_v1_profiles__profile_id__get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          },
          {
            "name": "__Host-session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Profile",
        "operationId": "patch_profile_api_v1_profiles__profile_id__patch",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "__Host-session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Profile",
        "operationId": "delete_profile_api_v1_profiles__profile_id__delete",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "__Host-session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Session"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles/{profile_id}/birth-time/normalize": {
      "post": {
        "summary": "Normalize Profile Birth Time",
        "operationId": "normalize_profile_birth_time_api_v1_profiles__profile_id__birth_time_normalize_post",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "__Host-session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NormalizeBirthTimeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthTimeNormalizationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BirthPlace": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "coordinate_source": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Coordinate Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "coordinate_source"
        ],
        "title": "BirthPlace"
      },
      "BirthTimeNormalizationResult": {
        "properties": {
          "original": {
            "$ref": "#/components/schemas/OriginalBirthRecord"
          },
          "historical_utc_offset_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Historical Utc Offset Minutes"
          },
          "dst_offset_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dst Offset Minutes"
          },
          "longitude_correction_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude Correction Minutes"
          },
          "equation_of_time_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equation Of Time Minutes"
          },
          "total_apparent_correction_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Apparent Correction Minutes"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/TimeCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "boundary_difference": {
            "$ref": "#/components/schemas/BoundaryDifference"
          },
          "correction_chain": {
            "items": {
              "$ref": "#/components/schemas/CorrectionStep"
            },
            "type": "array",
            "title": "Correction Chain"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "prohibited_conclusions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prohibited Conclusions"
          }
        },
        "type": "object",
        "required": [
          "original",
          "historical_utc_offset_minutes",
          "dst_offset_minutes",
          "longitude_correction_minutes",
          "equation_of_time_minutes",
          "total_apparent_correction_minutes",
          "candidates",
          "boundary_difference",
          "correction_chain",
          "warnings",
          "prohibited_conclusions"
        ],
        "title": "BirthTimeNormalizationResult"
      },
      "BoundaryDifference": {
        "properties": {
          "crosses_double_hour_boundary": {
            "type": "boolean",
            "title": "Crosses Double Hour Boundary"
          },
          "crosses_civil_date_boundary": {
            "type": "boolean",
            "title": "Crosses Civil Date Boundary"
          },
          "crosses_solar_term_boundary": {
            "type": "boolean",
            "title": "Crosses Solar Term Boundary"
          },
          "sensitive_rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sensitive Rules"
          }
        },
        "type": "object",
        "required": [
          "crosses_double_hour_boundary",
          "crosses_civil_date_boundary",
          "crosses_solar_term_boundary",
          "sensitive_rules"
        ],
        "title": "BoundaryDifference"
      },
      "CorrectionStep": {
        "properties": {
          "step": {
            "type": "string",
            "title": "Step"
          },
          "input_value": {
            "type": "string",
            "title": "Input Value"
          },
          "output_value": {
            "type": "string",
            "title": "Output Value"
          },
          "offset_minutes": {
            "type": "number",
            "title": "Offset Minutes"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          }
        },
        "type": "object",
        "required": [
          "step",
          "input_value",
          "output_value",
          "offset_minutes",
          "source_id"
        ],
        "title": "CorrectionStep"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvitationAccept": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 20,
            "title": "Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token"
        ],
        "title": "InvitationAccept"
      },
      "NormalizeBirthTimeRequest": {
        "properties": {
          "solar_term_instants_utc": {
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "type": "array",
            "title": "Solar Term Instants Utc",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "NormalizeBirthTimeRequest"
      },
      "OriginalBirthRecord": {
        "properties": {
          "calendar_type": {
            "type": "string",
            "enum": [
              "gregorian",
              "lunar"
            ],
            "title": "Calendar Type"
          },
          "local_date": {
            "type": "string",
            "format": "date",
            "title": "Local Date"
          },
          "local_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Time"
          },
          "timezone_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Timezone Id"
          },
          "timezone_database": {
            "type": "string",
            "title": "Timezone Database",
            "default": "IANA"
          },
          "timezone_database_version": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Timezone Database Version"
          },
          "time_precision": {
            "type": "string",
            "enum": [
              "minute",
              "double_hour",
              "half_day",
              "unknown"
            ],
            "title": "Time Precision"
          },
          "place": {
            "$ref": "#/components/schemas/BirthPlace"
          },
          "user_confirmed": {
            "type": "boolean",
            "title": "User Confirmed"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "calendar_type",
          "local_date",
          "local_time",
          "timezone_id",
          "timezone_database_version",
          "time_precision",
          "place",
          "user_confirmed",
          "captured_at"
        ],
        "title": "OriginalBirthRecord"
      },
      "ProfileCreate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "consent_version": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Consent Version"
          },
          "birth": {
            "$ref": "#/components/schemas/OriginalBirthRecord"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "consent_version",
          "birth"
        ],
        "title": "ProfileCreate"
      },
      "ProfilePatch": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "birth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OriginalBirthRecord"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProfilePatch"
      },
      "ProfileView": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "consent_version": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Consent Version"
          },
          "birth": {
            "$ref": "#/components/schemas/OriginalBirthRecord"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "consent_version",
          "birth",
          "id",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "ProfileView"
      },
      "SessionView": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "member",
              "viewer"
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "role"
        ],
        "title": "SessionView"
      },
      "TimeCandidate": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "basis": {
            "type": "string",
            "enum": [
              "civil",
              "local_mean_solar",
              "local_apparent_solar",
              "unknown_interval"
            ],
            "title": "Basis"
          },
          "local_datetime": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Datetime"
          },
          "utc_datetime": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utc Datetime"
          },
          "method_id": {
            "type": "string",
            "title": "Method Id"
          },
          "is_primary_chart": {
            "type": "boolean",
            "title": "Is Primary Chart",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "basis",
          "local_datetime",
          "utc_datetime",
          "method_id"
        ],
        "title": "TimeCandidate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}